moved filtering initialization to a separate function.
git-svn-id: svn+ssh://svn.r-forge.r-project.org/svnroot/genabel/pkg/filevector@346 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ fvlib/AbstractMatrix.h
@@ -69,8 +69,8 @@ public:
        delete [] ret;
     }
 
-    virtual unsigned int getNumVariables() = 0;
-    virtual unsigned int getNumObservations() = 0;
+    virtual unsigned long getNumVariables() = 0;
+    virtual unsigned long getNumObservations() = 0;
 
 	virtual void saveAs( string newFilename ) = 0;
 	virtual void saveVariablesAs( string newFilename, unsigned long nvars, unsigned long * varindexes) = 0;
